+++ context.go
@@ -180,13 +180,16 @@ func (ctx *Context) RunActionDirect(actionType NodeActionType, actionParam NodeA
 	return actionDetail, err
 }
 
-func (ctx *Context) overridePipeline(override string) bool {
-	return native.MaaContextOverridePipeline(ctx.handle, override)
+func (ctx *Context) overridePipeline(override string) error {
+	if !native.MaaContextOverridePipeline(ctx.handle, override) {
+		return errors.New("failed to override pipeline")
+	}
+	return nil
 }
 
 // OverridePipeline overrides pipeline.
 // The `override` parameter can be a JSON string or any data type that can be marshaled to JSON.
-func (ctx *Context) OverridePipeline(override any) bool {
+func (ctx *Context) OverridePipeline(override any) error {
 	switch v := override.(type) {
 	case string:
 		return ctx.overridePipeline(v)
@@ -199,14 +202,14 @@ func (ctx *Context) OverridePipeline(override any) bool {
 
 		jsonBytes, err := json.Marshal(v)
 		if err != nil {
-			return false
+			return err
 		}
 		return ctx.overridePipeline(string(jsonBytes))
 	}
 }
 
 // OverrideNext overrides the next list of task by name.
-func (ctx *Context) OverrideNext(name string, nextList []string) bool {
+func (ctx *Context) OverrideNext(name string, nextList []string) error {
 	list := buffer.NewStringListBuffer()
 	defer list.Destroy()
 	size := len(nextList)
@@ -221,32 +224,41 @@ func (ctx *Context) OverrideNext(name string, nextList []string) bool {
 			item.Destroy()
 		}
 	}()
-	return native.MaaContextOverrideNext(ctx.handle, name, list.Handle())
+	if !native.MaaContextOverrideNext(ctx.handle, name, list.Handle()) {
+		return errors.New("failed to override next")
+	}
+	return nil
 }
 
-func (ctx *Context) OverrideImage(imageName string, image image.Image) bool {
+func (ctx *Context) OverrideImage(imageName string, image image.Image) error {
 	img := buffer.NewImageBuffer()
 	defer img.Destroy()
 	img.Set(image)
-	return native.MaaContextOverrideImage(ctx.handle, imageName, img.Handle())
+	if !native.MaaContextOverrideImage(ctx.handle, imageName, img.Handle()) {
+		return errors.New("failed to override image")
+	}
+	return nil
 }
 
 // GetNodeJSON gets the node JSON by name.
-func (ctx *Context) GetNodeJSON(name string) (string, bool) {
+func (ctx *Context) GetNodeJSON(name string) (string, error) {
 	buf := buffer.NewStringBuffer()
 	defer buf.Destroy()
 	ok := native.MaaResourceGetNodeData(ctx.handle, name, buf.Handle())
-	return buf.Get(), ok
+	if !ok {
+		return "", errors.New("failed to get node JSON")
+	}
+	return buf.Get(), nil
 }
 
 func (ctx *Context) GetNodeData(name string) (*Node, error) {
-	raw, ok := ctx.GetNodeJSON(name)
-	if !ok {
-		return nil, errors.New("node not found")
+	raw, err := ctx.GetNodeJSON(name)
+	if err != nil {
+		return nil, err
 	}
 
 	var node Node
-	err := json.Unmarshal([]byte(raw), &node)
+	err = json.Unmarshal([]byte(raw), &node)
 	if err != nil {
 		return nil, err
 	}
@@ -273,26 +285,38 @@ func (ctx *Context) Clone() *Context {
 }
 
 // SetAnchor sets an anchor by name.
-func (ctx *Context) SetAnchor(anchorName, nodeName string) bool {
-	return native.MaaContextSetAnchor(ctx.handle, anchorName, nodeName)
+func (ctx *Context) SetAnchor(anchorName, nodeName string) error {
+	if !native.MaaContextSetAnchor(ctx.handle, anchorName, nodeName) {
+		return errors.New("failed to set anchor")
+	}
+	return nil
 }
 
 // GetAnchor gets an anchor by name.
-func (ctx *Context) GetAnchor(anchorName string) (string, bool) {
+func (ctx *Context) GetAnchor(anchorName string) (string, error) {
 	buf := buffer.NewStringBuffer()
 	defer buf.Destroy()
 	ok := native.MaaContextGetAnchor(ctx.handle, anchorName, buf.Handle())
-	return buf.Get(), ok
+	if !ok {
+		return "", errors.New("failed to get anchor")
+	}
+	return buf.Get(), nil
 }
 
 // GetHitCount gets the hit count of a node by name.
-func (ctx *Context) GetHitCount(nodeName string) (uint64, bool) {
+func (ctx *Context) GetHitCount(nodeName string) (uint64, error) {
 	var count uint64
 	ok := native.MaaContextGetHitCount(ctx.handle, nodeName, &count)
-	return count, ok
+	if !ok {
+		return 0, errors.New("failed to get hit count")
+	}
+	return count, nil
 }
 
 // ClearHitCount clears the hit count of a node by name.
-func (ctx *Context) ClearHitCount(nodeName string) bool {
-	return native.MaaContextClearHitCount(ctx.handle, nodeName)
+func (ctx *Context) ClearHitCount(nodeName string) error {
+	if !native.MaaContextClearHitCount(ctx.handle, nodeName) {
+		return errors.New("failed to clear hit count")
+	}
+	return nil
 }